refactor: extract PlayerPanel component
