+++ packages/web/src/Game.tsx
@@ -5,17 +5,15 @@ import {
   Resource,
   PopulationRole,
   RESOURCES,
-  STATS,
   POPULATION_ROLES,
   ACTION_INFO as actionInfo,
   DEVELOPMENT_INFO as developmentInfo,
-  LAND_ICON as landIcon,
-  SLOT_ICON as slotIcon,
   BUILDING_INFO as buildingInfo,
 } from '@kingdom-builder/engine';
-import type { EngineContext, ResourceKey } from '@kingdom-builder/engine';
+import type { ResourceKey } from '@kingdom-builder/engine';
 import { summarizeContent, describeContent, type Summary } from './translation';
 import { GameProvider, useGameEngine } from './state/GameContext';
+import PlayerPanel from './components/player/PlayerPanel';
 
 interface Action {
   id: string;
@@ -46,251 +44,6 @@ function renderSummary(summary: Summary | undefined): React.ReactNode {
     ),
   );
 }
-
-interface PlayerPanelProps {
-  player: EngineContext['activePlayer'];
-  className?: string;
-  buildingDescriptions: Map<string, Summary>;
-}
-
-const PlayerPanel: React.FC<PlayerPanelProps> = ({
-  player,
-  className = '',
-  buildingDescriptions,
-}) => {
-  const { ctx, handleHoverCard, clearHoverCard } = useGameEngine();
-  const popEntries = Object.entries(player.population).filter(([, v]) => v > 0);
-  const currentPop = popEntries.reduce((sum, [, v]) => sum + v, 0);
-  const popDetails = popEntries.map(([role, count]) => ({ role, count }));
-  function formatStatValue(key: string, value: number) {
-    if (key === 'absorption') return `${value * 100}%`;
-    return String(value);
-  }
-  const showPopulationCard = () =>
-    handleHoverCard({
-      title: '👥 Population',
-      effects: Object.values(POPULATION_ROLES).map(
-        (r) => `${r.icon} ${r.label} - ${r.description}`,
-      ),
-      effectsTitle: 'Archetypes',
-      requirements: [],
-      description:
-        'Population represents the people of your kingdom. Manage them wisely and assign roles to benefit your realm.',
-      bgClass: 'bg-gray-100 dark:bg-gray-700',
-    });
-
-  return (
-    <div className={`h-full flex flex-col space-y-1 ${className}`}>
-      <h3 className="font-semibold">{player.name}</h3>
-      <div className="panel-card flex flex-wrap items-center gap-2 px-3 py-2 w-fit">
-        {Object.entries(player.resources).map(([k, v]) => {
-          const info = RESOURCES[k as keyof typeof RESOURCES];
-          return (
-            <span
-              key={k}
-              className="bar-item hoverable cursor-help rounded px-1"
-              onMouseEnter={() =>
-                handleHoverCard({
-                  title: `${info.icon} ${info.label}`,
-                  effects: [],
-                  requirements: [],
-                  description: info.description,
-                  bgClass: 'bg-gray-100 dark:bg-gray-700',
-                })
-              }
-              onMouseLeave={clearHoverCard}
-            >
-              {info.icon}
-              {v}
-            </span>
-          );
-        })}
-        <div className="h-4 border-l border-black/10 dark:border-white/10" />
-        <span
-          className="bar-item hoverable cursor-help rounded px-1"
-          onMouseEnter={showPopulationCard}
-          onMouseLeave={clearHoverCard}
-        >
-          👥{currentPop}/{player.maxPopulation}
-          {popDetails.length > 0 && (
-            <>
-              {' ('}
-              {popDetails.map(({ role, count }, i) => {
-                const info =
-                  POPULATION_ROLES[role as keyof typeof POPULATION_ROLES];
-                return (
-                  <React.Fragment key={role}>
-                    {i > 0 && ','}
-                    <span
-                      className="cursor-help hoverable rounded px-1"
-                      onMouseEnter={(e) => {
-                        e.stopPropagation();
-                        handleHoverCard({
-                          title: `${info.icon} ${info.label}`,
-                          effects: [],
-                          requirements: [],
-                          description: info.description,
-                          bgClass: 'bg-gray-100 dark:bg-gray-700',
-                        });
-                      }}
-                      onMouseLeave={(e) => {
-                        e.stopPropagation();
-                        showPopulationCard();
-                      }}
-                    >
-                      {info.icon}
-                      {count}
-                    </span>
-                  </React.Fragment>
-                );
-              })}
-              {')'}
-            </>
-          )}
-        </span>
-        {Object.entries(player.stats)
-          .filter(([k]) => k !== 'maxPopulation')
-          .map(([k, v]) => {
-            const info = STATS[k as keyof typeof STATS];
-            return (
-              <span
-                key={k}
-                className="bar-item hoverable cursor-help rounded px-1"
-                onMouseEnter={() =>
-                  handleHoverCard({
-                    title: `${info.icon} ${info.label}`,
-                    effects: [],
-                    requirements: [],
-                    description: info.description,
-                    bgClass: 'bg-gray-100 dark:bg-gray-700',
-                  })
-                }
-                onMouseLeave={clearHoverCard}
-              >
-                {info.icon}
-                {formatStatValue(k, v)}
-              </span>
-            );
-          })}
-      </div>
-      {player.lands.length > 0 && (
-        <div className="flex flex-wrap gap-2 mt-2 w-fit">
-          {player.lands.map((land, idx) => {
-            const showLandCard = () =>
-              handleHoverCard({
-                title: `${landIcon} Land`,
-                effects: describeContent('land', land, ctx),
-                requirements: [],
-                effectsTitle: 'Developments',
-                bgClass: 'bg-gray-100 dark:bg-gray-700',
-              });
-            return (
-              <div
-                key={idx}
-                className="relative panel-card p-2 text-center hoverable cursor-help"
-                onMouseEnter={showLandCard}
-                onMouseLeave={clearHoverCard}
-              >
-                <span className="font-medium">{landIcon} Land</span>
-                <div className="mt-1 flex flex-wrap justify-center gap-1">
-                  {Array.from({ length: land.slotsMax }).map((_, i) => {
-                    const devId = land.developments[i];
-                    if (devId) {
-                      const name = ctx.developments.get(devId)?.name || devId;
-                      const title = `${developmentInfo[devId]?.icon || ''} ${name}`;
-                      const handleLeave = () => showLandCard();
-                      return (
-                        <span
-                          key={i}
-                          className="panel-card p-1 text-xs hoverable cursor-help"
-                          onMouseEnter={(e) => {
-                            e.stopPropagation();
-                            handleHoverCard({
-                              title,
-                              effects: describeContent(
-                                'development',
-                                devId,
-                                ctx,
-                                {
-                                  installed: true,
-                                },
-                              ),
-                              requirements: [],
-                              bgClass: 'bg-gray-100 dark:bg-gray-700',
-                            });
-                          }}
-                          onMouseLeave={(e) => {
-                            e.stopPropagation();
-                            handleLeave();
-                          }}
-                        >
-                          {developmentInfo[devId]?.icon} {name}
-                        </span>
-                      );
-                    }
-                    const handleLeave = () => showLandCard();
-                    return (
-                      <span
-                        key={i}
-                        className="panel-card p-1 text-xs hoverable cursor-help whitespace-nowrap"
-                        onMouseEnter={(e) => {
-                          e.stopPropagation();
-                          handleHoverCard({
-                            title: `${slotIcon} Development Slot (empty)`,
-                            effects: [],
-                            description: `Use ${actionInfo['develop']?.icon ?? ''} ${actionInfo['develop']?.label ?? ''} to build here`,
-                            requirements: [],
-                            bgClass: 'bg-gray-100 dark:bg-gray-700',
-                          });
-                        }}
-                        onMouseLeave={(e) => {
-                          e.stopPropagation();
-                          handleLeave();
-                        }}
-                      >
-                        {slotIcon} Development Slot (empty)
-                      </span>
-                    );
-                  })}
-                </div>
-              </div>
-            );
-          })}
-        </div>
-      )}
-      {player.buildings.size > 0 && (
-        <div className="flex flex-wrap gap-2 mt-2 w-fit">
-          {Array.from(player.buildings).map((b) => {
-            const name = ctx.buildings.get(b)?.name || b;
-            const icon =
-              buildingInfo[b]?.icon || actionInfo['build']?.icon || '';
-            const title = `${icon} ${name}`;
-            return (
-              <div
-                key={b}
-                className="panel-card p-2 text-center hoverable cursor-help"
-                onMouseEnter={() =>
-                  handleHoverCard({
-                    title,
-                    effects: buildingDescriptions.get(b) ?? [],
-                    requirements: [],
-                    bgClass: 'bg-gray-100 dark:bg-gray-700',
-                  })
-                }
-                onMouseLeave={clearHoverCard}
-              >
-                <span className="font-medium">
-                  {icon} {name}
-                </span>
-              </div>
-            );
-          })}
-        </div>
-      )}
-    </div>
-  );
-};
-/* eslint-enable @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-unnecessary-type-assertion */
 
 function renderCosts(
   costs: Record<string, number> | undefined,
